started on unit testing for first status button
