+++ package.json
@@ -36,7 +36,11 @@
                 "testy.testGeneration.setTestFileNameGenerationType": {
                     "type": "string",
                     "default": "prefix",
-                    "enum": ["prefix", "suffix", "none"],
+                    "enum": [
+                        "prefix",
+                        "suffix",
+                        "none"
+                    ],
                     "enumDescriptions": [
                         "Adds a prefix to the start of a file name. Example: (prefix 'test_' for python tests) test_class_file.py.",
                         "Adds a suffix to the end of a file name. Example:(suffix 'Test' for java tests) ClassFileTest.java.",
@@ -63,19 +67,21 @@
         "test": "node ./out/test/runTest.js"
     },
     "devDependencies": {
-        "@types/vscode": "^1.73.0",
         "@types/glob": "^8.0.0",
         "@types/mocha": "^10.0.1",
         "@types/node": "16.x",
+        "@types/sinon": "^10.0.13",
+        "@types/vscode": "^1.73.0",
         "@typescript-eslint/eslint-plugin": "^5.45.0",
         "@typescript-eslint/parser": "^5.45.0",
+        "@vscode/test-electron": "^2.2.0",
         "eslint": "^8.28.0",
         "glob": "^8.0.3",
         "mocha": "^10.1.0",
-        "typescript": "^4.9.3",
+        "sinon": "^15.0.0",
         "ts-loader": "^9.4.1",
+        "typescript": "^4.9.3",
         "webpack": "^5.75.0",
-        "webpack-cli": "^5.0.0",
-        "@vscode/test-electron": "^2.2.0"
+        "webpack-cli": "^5.0.0"
     }
 }
