+++ src/test/unit/statusItems/generateTestStatusItem.test.ts
@@ -1,5 +1,16 @@
+import * as vscode from 'vscode';
+import * as sinon from 'sinon';
+import GenerateTestStatusItem from "../../../statusItems/generateTestStatusItem";
+
 describe('Generate Test File', () => {
     describe('Generate test file name', () => {
+        let generateTestStatusItem: GenerateTestStatusItem;
+        let context: sinon.SinonStubbedInstance<vscode.ExtensionContext>;
+
+        beforeEach(() => {
+            generateTestStatusItem = new GenerateTestStatusItem(context);
+        });
+
         describe('Generate test file with custom prefix', () => {
             
         });
